requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
pytest
pymupdf
python-docx
textblob
redis
openai
pytesseract
fpdf2